Initial commit.  Partially working application.
This is the initial commit.  Implements an index.html file and an
app.js angular file.  The app.js file implements an angular service
and one angular controller.  There is still much refactoring and
additional work to do.  For instance, need to implement directives
to isolate scope so not all inputs are sharing the parent scope.
Much to be done, but this is a good start.