Add directive to isolate scope for purchase inputs.
This commit adds a directive element to abstract the 'add purchase'
functionality for each atm withdrawal.  This enables scope isolation
from the parent controller.  Meaning that when an amount is input for a
new purchase for a particulary atm withdrawal, that same amount is not
broadcast to the rest of the atm withdrawals--it's isolated to only that
withdrawal.  Also made a few minor refactoring edits and set up some
scaffolding for the next iteration.