Retrieve dummy data from external json file.
This commit adds the functionality to retrieve initial dummy data from
an external json file.  It implements a resolve attribute on the route,
which calls on the service to asynchronously call on the $http service
to get the data from the file.  Once the resolved promise is returned,
then the controller can be instantiated and render the view.

Still need to set up a watch on the service to update the controller
when a new purchase is added (to the service).  Also, thinking of
revising the view--thinking of a list on the left and only one
'detailed' view in the middle.