Fix bug to add purchase to an atm withdrawal.
This commit fixes a bug which now allows a purchase to be added to a
particulary atm withdrawal.  When a purchase is added, it is pushed to
the 'purchases' property of the atm withdrawal, which is implemented as
an arrary of purchase object literals.  Once added, the purchase list is
updated (for that withdrawal only) and the 'total spent' and 'amount
remaining' are recalculated to reflect the additional purchase.  No
error checking is in place yet--that will be next along with converting
the date to a Date obj.