Change date representation in data.json
Changes the date representation in the json data file since the json
spec does not allow for objects as values.  Change the date to be
represented by and obj literal with {"year": xxxx, "month": xx, "day":
xx} properties.  This should be simple to create the javascript date
objects when the json file is read into memory.