Create js date obj from date literal in json file.
This commit adds functionality to create an actual js date object to
store in the in-memory representation of the atm withdrawals.  JS
objects are not allowed to be stored as values in a json file;
therefore, the initial date json file represented the date of the atm
withdrawal as an obj literal with all the required information.  This
commit uses that obj literal to create a js date obj to store in the
in-memory array.  This makes it much easier to represent the date in the
templates, especially when it is piped through the date filter.