Add basic structure for ATM sidebar list.
This commit adds the basic structure and functionality for the ATM
sidebar list group.  It adds an additonal controller and modifies the
route slightly to include the id paramater in the URL.  The data is
still hard coded in the controller for testing purposes, but will be
abstracted out as the iterative development continues.